requirements.txt
# 필수 라이브러리
maxminddb==1.5.4
redis==2.10.5
tornado==4.4.2
tornado-redis==2.4.18
python-dotenv==0.21.1
scapy==2.6.1

# 추가 유틸리티
requests==2.28.2
pytz==2023.3
